Moved index file and fix imports
