+++ modules/client/examples/02-tokenVoting-client/01-installation.ts
@@ -14,7 +14,7 @@ import {
   TokenVotingClient,
   VotingMode
 } from "@aragon/sdk-client";
-import { context } from "../01-client/index";
+import { context } from "../index";
 
 // Insantiate the general purpose client from the Aragon OSx SDK context.
 const client: Client = new Client(context);
