+++ modules/client/examples/02-tokenVoting-client/02-create-proposal.ts
@@ -16,7 +16,7 @@ import {
   VotingSettings,
   VoteValues
 } from "@aragon/sdk-client";
-import { context } from "../01-client/index";
+import { context } from "../index";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
