+++ modules/client/examples/02-tokenVoting-client/11-pin-metadata.ts
@@ -9,7 +9,7 @@ import {
   TokenVotingClient,
   ProposalMetadata
 } from "@aragon/sdk-client";
-import { context } from "../01-client/index";
+import { context } from "../index";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
