+++ modules/client/examples/03-multisig-client/01-installation.ts
@@ -17,7 +17,7 @@ import {
   MultisigClient,
   MultisigPluginInstallParams
 } from "@aragon/sdk-client";
-import { context } from "../01-client/index";
+import { context } from "../index";
 
 // Instantiate a client from the Aragon OSx SDK context.
 const client: Client = new Client(context);
