+++ modules/client/examples/03-multisig-client/02-create-proposal.ts
@@ -14,7 +14,7 @@ import {
   TokenType,
   WithdrawParams
 } from "@aragon/sdk-client";
-import { context } from "../01-client/index";
+import { context } from "../index";
 
 // Instantiate an Aragon OSx SDK client.
 const client: Client = new Client(context);
