+++ modules/client/examples/04-addresslistVoting-client/02-create-proposal.ts
@@ -15,7 +15,7 @@ import {
   VotingMode,
   VoteValues
 } from "@aragon/sdk-client";
-import { context } from "../01-client/index";
+import { context } from "../index";
 
 // Create a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
