+++ modules/client/examples/05-encoders/04-register-standard-callback.ts
@@ -5,7 +5,7 @@ Encodes the action of registering a new standard callback for the DAO.
 */
 
 import { Client, ContextPlugin, DaoAction } from "@aragon/sdk-client";
-import { context } from '../01-client/index';
+import { context } from '../index';
 
 // Initialize the plugin's context from the Aragon SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
