+++ modules/client/examples/06-decoders/03-withdraw.ts
@@ -5,7 +5,7 @@ Decodes the parameters of a withdraw action of any token type.
 */
 
 import { Client, WithdrawParams } from "@aragon/sdk-client";
-import { context } from "../01-client/index";
+import { context } from "../index";
 
 // Insantiates an Aragon OSx SDK client.
 const client: Client = new Client(context);
